# docs/api/project-settings.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Project settings — review S03/S10",
    "version": "0.10.1"
  },
  "paths": {
    "/api/v1/projects": {
      "parameters": [],
      "get": {
        "operationId": "listProjects",
        "description": "",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createProject",
        "description": "Chỉ ADMIN tạo dự án; tự có membership PM.",
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProject"
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}": {
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        }
      ],
      "get": {
        "operationId": "getProject",
        "description": "",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateProject",
        "description": "",
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProject"
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/members": {
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        }
      ],
      "get": {
        "operationId": "listProjectMembers",
        "description": "",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Member"
                  }
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/members/{userId}": {
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        },
        {
          "name": "userId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "maxLength": 36
          }
        }
      ],
      "put": {
        "operationId": "setProjectMember",
        "description": "expectedVersion chỉ null khi chưa từng có membership. Không đổi quyền hệ thống.",
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Member"
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetMember"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "removeProjectMember",
        "description": "Gỡ mềm; chặn PM cuối.",
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "responses": {
          "204": {
            "description": "Đã lưu trữ hoặc gỡ thành viên"
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "expectedVersion",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ]
      }
    },
    "/api/v1/projects/{projectId}/member-candidate": {
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        }
      ],
      "get": {
        "operationId": "findMemberCandidate",
        "description": "Chỉ PM hoặc ADMIN là thành viên; tra username chính xác, không cấp danh sách tài khoản.",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberCandidate"
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "TMS_SESSION"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-TOKEN"
      }
    },
    "schemas": {
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "code": {
            "type": "string",
            "maxLength": 32
          },
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          },
          "timezone": {
            "type": "string",
            "maxLength": 50
          },
          "version": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "archived": {
            "type": "boolean"
          },
          "projectRole": {
            "type": "string",
            "maxLength": 16
          }
        },
        "required": []
      },
      "CreateProject": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 32,
            "pattern": "[A-Za-z0-9][A-Za-z0-9_-]{1,31}"
          },
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          },
          "timezone": {
            "type": "string",
            "maxLength": 50
          }
        },
        "required": [
          "code",
          "name",
          "timezone"
        ]
      },
      "UpdateProject": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          },
          "timezone": {
            "type": "string",
            "maxLength": 50
          },
          "expectedVersion": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        },
        "required": [
          "expectedVersion"
        ]
      },
      "Member": {
        "type": "object",
        "properties": {
          "membershipId": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "userId": {
            "type": "string",
            "maxLength": 36
          },
          "username": {
            "type": "string",
            "maxLength": 64
          },
          "displayName": {
            "type": "string",
            "maxLength": 100
          },
          "systemRole": {
            "type": "string",
            "maxLength": 16
          },
          "projectRole": {
            "enum": [
              "PM",
              "TESTER",
              "MEMBER"
            ],
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "version": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        },
        "required": []
      },
      "SetMember": {
        "type": "object",
        "properties": {
          "projectRole": {
            "type": "string",
            "enum": [
              "PM",
              "TESTER",
              "MEMBER"
            ]
          },
          "expectedVersion": {
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "nullable": true
          }
        },
        "required": [
          "projectRole"
        ]
      },
      "MemberCandidate": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "maxLength": 36
          },
          "username": {
            "type": "string",
            "maxLength": 64
          },
          "displayName": {
            "type": "string",
            "maxLength": 100
          },
          "membershipVersion": {
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "nullable": true
          }
        },
        "required": []
      },
...
      "ProjectDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "code": {
            "type": "string",
            "maxLength": 32
          },
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          },
          "timezone": {
            "type": "string",
            "maxLength": 50
          },
          "version": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "archived": {
            "type": "boolean"
          },
          "membersCount": {
            "type": "integer",
            "minimum": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedBy": {
            "type": "string"
          }
        }
      }
    }
  }
}